requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
httpx==0.28.1
aiosqlite==0.20.0
pyyaml==6.0.2
jinja2==3.1.5

# PaddleOCR
paddlepaddle>=2.6.0
paddleocr>=2.7.0

# PDF to image
PyMuPDF>=1.23.0
